Initial commit of ace-archive keeper client tool.
Add basic download functionality for ace archive keeper client tool.
Fetch a list of all the artifacts in the archive from the ace archive
API. For each file in each artifact check if any of the files already
exist in the specified download directory. If they do, check if their
checksum matches the checksum returned by the API. If it does, proceed
to the next file. If it doesn't download the new file. Once all files
have been downloaded, check if the on-disk metadata.json file for this
artifact matches the metadata returned by the api. If it doesn't update
the metadata.json.

Keep a running hash of all the metadata of all artifacts processed so
far. After an artifact's on-disk metadata has been updated (or not)
update the running hash. Once all artifacts are backed up, log the
final archive hash.