Keep a running tally of file sizes when processing
After completing a file download, or after confirming a download is not
required by verifying the checksum, read the size of the file. Keep a
running tally inside the artifact of it's total file size.

When all files in an artifact have been processed, add that artifact's
total size to a running tally for the total archive size.