Store backed-up archive as an uncompressed zip.
Store the backed-up archive as an uncompressed zip file instead of a
directory tree. This prevents accidental modification of the files in
the backup. Using a zip without compression makes extracting &
recreating the zip less CPU intensive.

When performing a backup, extract the zip to a temporary directory to
perform checksum comparisons on the files. After fetching any updated
files & writing up-to-date file metadata create a new zip file from the
contents of the tempdir and write that file over previous zip. This
provides some guards against accidental deletions of files by this
script.

Move any unextractable file provided for the --archive-zip argument to
prevent it from being overwritten.