Prune deleted artifacts & files from backup zip
Handle unexpected file pruning in two ways.

1. Prune deleted files from known artifacts.
After completing the backup phase for each artifact, the filetree
starting at that artifact's dir is walked and any files or folders that
are not in the artifact's 'files' field are deleted. Symlinks are
not followed for safety.

2. Prune unknown artifacts.
After completing the backup phase for all artifacts, the top-level
directory used for the backup is scanned. Any files or dirs that don't
correspond to a known artifact ID are recursively deleted. Symlinks are
not followed for safety.

For added safety, decline to prune files if the directory to be pruned
is outside of the system temp. This is intended to prevent accidentally
deleting files not managed by the keeper script.