Use new raw_url field for file fetching.
Instead of downloading from the url field on files, use the raw_url
field, which is always the url to the actual backend file.

The url in the url field uses content negotiation to decide if the
content should be rendered in an iframe with additional navigation links
or not. This works for PDF files, but fails for HTML documents. So use
the new raw_url field instead.

Update the Artifact & ArtifactFile dataclasses to include the new fields
returned by the API.

Update the checksum generating portion of the keeper server to expect
the new fields.