Detect artifact renames & restructure zip
If an unexpected dir exists in the artifacts dir of an incremental
backup, check if it contains artifact metadata with an artifact ID.
If that ID matches a valid artifact, move the unexpected dir to that
artifact's new location and start an incremental backup.

Restructure the zip file to place artifact dirs under an 'artifacts'
dir. Use the artifact slug instead of the artifact ID for the artifact
dir name.