Add work-in-progress tkinter based gui
Use tkinter to build a GUI frontend for acearchive keeper.
The GUI has basic options as the CLI. Advanced options are omitted.

When the backup is started a new worker process is spawned to perform
the backup. This needs to be a separate process and can't simply be a
separate thread as the GIL will cause the GUI will become non-responsive
during CPU-bound. Communication between the worker process and the GUI
is done via multiprocessing queues.

When canceled or closed the worker is asked to cancel and clean itself
up. This can take a few seconds. We don't want to simply terminate the
worker process as this will prevent the cleanup code from running, and
any tempdirs created will not be deleted.

Create tasks for each artifact backup instead of simply gathering them
so we can count up the number of completed artifacts so far. Use this
for reporting a progress % to the GUI.

Move the main logic from __init__.py to worker.py to prevent a potential
circular import if using __init__.py as the GUI entrypoint.

Fix config file sometimes using None or False when no option is
specified instead of the empty string for string default values.

Adjust how logging handlers created and inherited to prevent
accidentally crashing the GUI by logging to a logger owned by another
thread. Also explicitly use the root logger in all entrypoints to ensure
proper handler inheritance.