Use pydantic validated toml instead of ini config.
Replace configparser with more flexible TOML config file format. Use
pydantic models to validate read from and written to the config file.

For the CLI, only write the keeper ID to the config file. The config
file is required as the keeper ID is a required bit of data and that is
stored in the config file. Config file usage needs to be re-thought.
Because the config file is non-optional, any fields included in the
config file can't be overridden by not specifying them at the CLI.

For example, if email is present in the config file there is no CLI
option to disable its usage. As such, only read the keeper ID from the
config file for now until the config file can be made optional. This may
require re-locating the keeper ID.